Binary search algorithm in php
steps:-
i)  Binary search works with only sorted array.
ii) The time complexity of a binary search is O(logn).
iii) In binary search, we first calculate the mid value. After that we compare value needs to search to the middle element of an array. If a value matches with the middle element of an array then it’s index is returned otherwise the algorithm repeats if the value is less than or greater than the middle element.