Created Language-Translator.py
we are using Google API to build a Language Translator which can translate one language to another language. On the internet, we can see lots of projects on Speech Recognition, Speech to text, text to speech, etc. but here in this project we are going to build something more advance than that.