Add files via upload
 A palindrome is a phrase that reads the same forward and backward
(examples: ‘racecar’, ‘radar’, ‘noon’, or ‘rats live on no evil star’). By extension we call
every string a palindrome that reads the same from left to right and from right to left.
Develop a recursive algorithm that takes as input a string and decides whether the
string is a palindrome. Write down your algorithm in pseudocode.