ReflexTime
This page uses html, CSS and JS to create a  code that is used to test he reaction time of the user, by measuring the time between  a colored div appears and the clicking of those divs by the user. this measurement of time uses javascript. This game is never ending and can be played for as long as the user wishes. Hope you like it.....